+++ cli/src/commands/git/push.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::collections::HashMap;
 use std::collections::HashSet;
 use std::fmt;
 use std::io;
@@ -19,8 +20,11 @@ use std::io::Write;
 
 use clap::ArgGroup;
 use clap_complete::ArgValueCandidates;
+use indexmap::IndexSet;
 use itertools::Itertools;
 use jj_lib::backend::CommitId;
+use jj_lib::commit::Commit;
+use jj_lib::commit::CommitIteratorExt as _;
 use jj_lib::config::ConfigGetResultExt as _;
 use jj_lib::git;
 use jj_lib::git::GitBranchPushTargets;
@@ -34,6 +38,7 @@ use jj_lib::refs::LocalAndRemoteRef;
 use jj_lib::repo::Repo;
 use jj_lib::revset::RevsetExpression;
 use jj_lib::settings::UserSettings;
+use jj_lib::signing::SignBehavior;
 use jj_lib::str_util::StringPattern;
 use jj_lib::view::View;
 
@@ -295,7 +300,38 @@ pub fn cmd_git_push(
         return Ok(());
     }
 
-    validate_commits_ready_to_push(ui, &bookmark_updates, &remote, &tx, args)?;
+    let sign_behavior = if tx.settings().get_bool("git.sign-on-push")? {
+        Some(SignBehavior::Own)
+    } else {
+        None
+    };
+    let commits_to_sign =
+        validate_commits_ready_to_push(ui, &bookmark_updates, &remote, &tx, args, sign_behavior)?;
+    if !args.dry_run && !commits_to_sign.is_empty() {
+        if let Some(sign_behavior) = sign_behavior {
+            let num_updated_signatures = commits_to_sign.len();
+            let num_rebased_descendants;
+            (num_rebased_descendants, bookmark_updates) = sign_commits_before_push(
+                &mut tx,
+                commits_to_sign,
+                sign_behavior,
+                bookmark_updates,
+            )?;
+            if let Some(mut formatter) = ui.status_formatter() {
+                writeln!(
+                    formatter,
+                    "Updated signatures of {num_updated_signatures} commits"
+                )?;
+                if num_rebased_descendants > 0 {
+                    writeln!(
+                        formatter,
+                        "Rebased {num_rebased_descendants} descendant commits"
+                    )?;
+                }
+            }
+        }
+    }
+
     if let Some(mut formatter) = ui.status_formatter() {
         writeln!(formatter, "Changes to push to {remote}:")?;
         print_commits_ready_to_push(formatter.as_mut(), tx.repo(), &bookmark_updates)?;
@@ -335,14 +371,17 @@ pub fn cmd_git_push(
 }
 
 /// Validates that the commits that will be pushed are ready (have authorship
-/// information, are not conflicted, etc.)
+/// information, are not conflicted, etc.).
+///
+/// Returns the list of commits which need to be signed.
 fn validate_commits_ready_to_push(
     ui: &Ui,
     bookmark_updates: &[(String, BookmarkPushUpdate)],
     remote: &str,
     tx: &WorkspaceCommandTransaction,
     args: &GitPushArgs,
-) -> Result<(), CommandError> {
+    sign_behavior: Option<SignBehavior>,
+) -> Result<Vec<Commit>, CommandError> {
     let workspace_helper = tx.base_workspace_helper();
     let repo = workspace_helper.repo();
 
@@ -369,6 +408,13 @@ fn validate_commits_ready_to_push(
     } else {
         Box::new(|_: &CommitId| Ok(false))
     };
+    let sign_settings = sign_behavior.map(|sign_behavior| {
+        let mut sign_settings = settings.sign_settings();
+        sign_settings.behavior = sign_behavior;
+        sign_settings
+    });
+
+    let mut commits_to_sign = vec![];
 
     for commit in workspace_helper
         .attach_revset_evaluator(commits_to_push)
@@ -418,8 +464,61 @@ fn validate_commits_ready_to_push(
             }
             return Err(error);
         }
+        if let Some(sign_settings) = &sign_settings {
+            if !commit.is_signed() && sign_settings.should_sign(commit.store_commit()) {
+                commits_to_sign.push(commit);
+            }
+        }
     }
-    Ok(())
+    Ok(commits_to_sign)
+}
+
+/// Signs commits before pushing.
+///
+/// Returns the number of commits with rebased descendants and the updated list
+/// of bookmark names and corresponding [`BookmarkPushUpdate`]s.
+fn sign_commits_before_push(
+    tx: &mut WorkspaceCommandTransaction,
+    commits_to_sign: Vec<Commit>,
+    sign_behavior: SignBehavior,
+    bookmark_updates: Vec<(String, BookmarkPushUpdate)>,
+) -> Result<(usize, Vec<(String, BookmarkPushUpdate)>), CommandError> {
+    let commit_ids: IndexSet<CommitId> = commits_to_sign.iter().ids().cloned().collect();
+    let mut old_to_new_commits_map: HashMap<CommitId, CommitId> = HashMap::new();
+    let mut num_rebased_descendants = 0;
+    tx.repo_mut()
+        .transform_descendants(commit_ids.iter().cloned().collect_vec(), |rewriter| {
+            let old_commit_id = rewriter.old_commit().id().clone();
+            if commit_ids.contains(&old_commit_id) {
+                let commit = rewriter
+                    .reparent()
+                    .set_sign_behavior(sign_behavior)
+                    .write()?;
+                old_to_new_commits_map.insert(old_commit_id, commit.id().clone());
+            } else {
+                num_rebased_descendants += 1;
+                let commit = rewriter.reparent().write()?;
+                old_to_new_commits_map.insert(old_commit_id, commit.id().clone());
+            }
+            Ok(())
+        })?;
+
+    let bookmark_updates = bookmark_updates
+        .into_iter()
+        .map(|(bookmark_name, update)| {
+            (
+                bookmark_name,
+                BookmarkPushUpdate {
+                    old_target: update.old_target,
+                    new_target: update
+                        .new_target
+                        .map(|id| old_to_new_commits_map.get(&id).cloned().unwrap_or(id)),
+                },
+            )
+        })
+        .collect_vec();
+
+    Ok((num_rebased_descendants, bookmark_updates))
 }
 
 fn print_commits_ready_to_push(